requirements.txt
httpx
requests
python-dotenv
beautifulsoup4
langchain
langchain-google-genai
langchain-community
langchain-core
tiktoken
pytest
openai
pillow
streamlit
Markdown
matplotlib
google-genai
selenium
pydantic
google-cloud-texttospeech
moviepy==1.0.3